# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AfriGest Admin & Partners API (Skeleton)
  version: 0.1.0
servers:
  - url: https://api.example.com
paths:
...
  /inventory/summary:
    get:
      summary: Inventory summary for a boutique
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: boutiqueId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
  /inventory/sessions:
    post:
      summary: Create an inventory session and compute variance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                boutiqueId: { type: string }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      productId: { type: string }
                      counted: { type: number }
                      unitPrice: { type: number }
      responses:
        '201':
          description: Created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
...